requirements.txt
fastf1
pandas
fastparquet
plotly
solara
pillow
